+++ src/main/java/com/sparta/northwind/controllers/ShipperController.java
@@ -1,5 +1,6 @@
 package com.sparta.northwind.controllers;
 
+import com.sparta.northwind.entities.Customer;
 import com.sparta.northwind.entities.Shipper;
 import com.sparta.northwind.repositories.ShipperRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,13 +25,32 @@ public class ShipperController {
 		Optional<Shipper> optionalShipper = repository.findById(id);
 		return optionalShipper.map(shipper -> new ResponseEntity<>(shipper, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
 	}
-	@DeleteMapping("/shippers/delete/{id}")
+	@DeleteMapping("/shippers/{id}")
 	public ResponseEntity<Shipper> deleteShipper(@PathVariable int id){
-		repository.deleteById(id);
-		return new ResponseEntity<>(HttpStatus.OK);
+		if (repository.existsById(id)){
+			repository.deleteById(id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		}
+		else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
-	@PostMapping("/shippers/create")
-	public Shipper createShipper(@RequestBody Shipper shipper){
-		return repository.save(shipper);
+	@PostMapping("/shippers")
+	public ResponseEntity<Shipper> createShipper(@RequestBody Shipper shipper){
+		return new ResponseEntity<>(repository.save(shipper), HttpStatus.CREATED);
+	}
+
+	@PutMapping(value = "/shippers/{id}", consumes = {"application/json"})
+	public ResponseEntity<Shipper> updateEmployee(@PathVariable int id, @RequestBody Shipper shipper) {
+		Optional<Shipper> optionalShipper = repository.findById(id);
+
+		if (optionalShipper.isPresent()) {
+			shipper.setId(id);
+			repository.save(shipper);
+
+			return new ResponseEntity<>(shipper, HttpStatus.OK);
+		}
+		return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
 	}
 }
